[executor] merge block_processor into lib.rs
